requirements.txt
autocommand==2.2.2
backports.tarfile==1.2.0
blosc2==2.7.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.1.0
ndindex==1.9.2
networkx==3.4.2
numexpr==2.10.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.1
pandas==2.2.3
pillow==11.1.0
platformdirs==4.2.2
py-cpuinfo==9.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scipy==1.15.0
six==1.17.0
sympy==1.13.1
tables==3.10.1
tomli==2.0.1
torch==2.5.1
torchdiffeq==0.2.5
tqdm==4.67.1
triton==3.1.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.19.2
